# openapi-schema.yaml
openapi: 3.0.0
info:
  title: NaviDocs API
  description: Comprehensive boat documentation management API with 16 supporting tables
  version: 1.0.0
  contact:
    name: NaviDocs Team
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.navidocs.com/v1
    description: Production server
paths:
...
  /boats/{boatId}/inventory:
    get:
      summary: List inventory items for a boat
      operationId: listInventoryItems
      parameters:
        - name: boatId
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
    post:
      summary: Create a new inventory item
      operationId: createInventoryItem
      parameters:
        - name: boatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryItemRequest'
      responses:
        '201':
          description: Item created successfully
  /inventory/{itemId}:
    get:
      summary: Get a specific inventory item
      operationId: getInventoryItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
    put:
      summary: Update an inventory item
      operationId: updateInventoryItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryItemRequest'
      responses:
        '200':
          description: Item updated successfully
    delete:
      summary: Delete an inventory item
      operationId: deleteInventoryItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Item deleted successfully

  # Maintenance Records Endpoints
...
components:
  schemas:
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        boatId:
          type: integer
        name:
          type: string
        category:
          type: string
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
          format: decimal
        currentValue:
          type: number
          format: decimal
        photoUrls:
          type: array
          items:
            type: string
        depreciationRate:
          type: number
          format: decimal
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateInventoryItemRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        category:
          type: string
        purchaseDate:
          type: string
          format: date
        purchasePrice:
          type: number
          format: decimal
        currentValue:
          type: number
          format: decimal
        photoUrls:
          type: array
          items:
            type: string
        depreciationRate:
          type: number
          format: decimal
        notes:
          type: string

    # Maintenance Record Schema
...